Add files via upload
These are the screenshots of the program. Sign In shows the opening page. Admin Access shows the leading characters to bring up the admin gui pop up. Pop Up is the admin's access to get to the GUI or return to the sign in. tempTable is the part of the GUI that would show the table and allow the user to edit the database. However, my program is for some reason unable to authenticate properly with the database I am using(MongoDB). When the application attempts to perform any action on the database, it gives me an error message for invalid credentials. This happens no matter the credentials or permissions I use.